core,params: rename pkg oldmultigeth->multigethv0
