+++ params/types/genesisT/gen_genesis.go
@@ -31,7 +31,7 @@ import (
 	common0 "github.com/ethereum/go-ethereum/params/types/ctypes"
 	"github.com/ethereum/go-ethereum/params/types/goethereum"
 	"github.com/ethereum/go-ethereum/params/types/multigeth"
-	"github.com/ethereum/go-ethereum/params/types/oldmultigeth"
+	"github.com/ethereum/go-ethereum/params/types/multigethv0"
 )
 
 var _ = (*genesisSpecMarshaling)(nil)
@@ -101,8 +101,8 @@ func (g *Genesis) UnmarshalJSON(input []byte) error {
 	switch conf.(type) {
 	case *multigeth.MultiGethChainConfig:
 		dec.Config = &multigeth.MultiGethChainConfig{}
-	case *oldmultigeth.ChainConfig:
-		dec.Config = &oldmultigeth.ChainConfig{}
+	case *multigethv0.ChainConfig:
+		dec.Config = &multigethv0.ChainConfig{}
 	case *goethereum.ChainConfig:
 		dec.Config = &goethereum.ChainConfig{}
 	default:
