+++ params/types/multigethv0/oldmultigeth_chain_config.go
@@ -1,4 +1,4 @@
-package oldmultigeth
+package multigethv0
 
 import (
 	"math/big"
