+++ params/types/multigethv0/oldmultigeth_chain_config_configurator.go
@@ -1,4 +1,4 @@
-package oldmultigeth
+package multigethv0
 
 import (
 	"math/big"
